requirements.txt
google-generativeai
google-search-results

fastapi
uvicorn
dotenv 